requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
Werkzeug==3.0.3
gunicorn
lxml
numpy-financial 
shap
pyyaml
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
scikit-learn==1.6.1
scipy==1.15.2
lightgbm==4.6.0
scikit-optimize==0.10.2
joblib==1.4.2
python-dotenv==1.0.1

